Move deformable camera tasks to benchmark package
